pyproject.toml
[project]
name = "lunvex-code"
version = "0.1.0"
description = "AI coding assistant powered by DeepSeek-V3 - Build your own Claude Code"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Artem Melnik", email = "artemmelnik@example.com" },
]
keywords = [
    "ai",
    "coding",
    "assistant",
    "lunvex",
    "cli",
    "claude-code",
    "open-source",
]
homepage = "https://github.com/artemmelnik/lunvex-code"
repository = "https://github.com/artemmelnik/lunvex-code"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
]
dependencies = [
    "openai>=2.31.0",
    "typer>=0.24.1",
    "rich>=14.3.4",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.2",
    "tomli>=2.4.1",
    "tomli-w>=1.2.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "black>=26.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
    "pip-audit>=2.10.0",
]

[project.scripts]
lunvex-code = "lunvex_code.cli:main"
lvc = "lunvex_code.cli:main"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "lunvex_code",
    "deepseek_code",
]

[tool.black]
line-length = 100
target-version = [
    "py310",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "W",
]
ignore = [
    "E501",
]

[tool.ruff.format]
quote-style = "double"
